refactor: decompose core and memory crate oversized modules (#442)
Split oversized modules to stay under 300 LOC per file:

Core crate (client/):
- types.rs (678 LOC) -> types.rs (221) + request.rs (341) + response.rs (157)
- mod.rs (651 LOC) -> mod.rs (475) + retry.rs (63) + tool_loop.rs (135)

Memory crate:
- types.rs (464 LOC) -> types.rs (174) + entry.rs (185) + query.rs (135)

Public API unchanged -- all types re-exported from parent modules for
backward compatibility. All existing import paths continue to work.
