+++ crates/core/src/client/mod.rs
@@ -38,7 +38,11 @@
 
 pub mod config;
 pub mod error;
+pub mod request;
+pub mod response;
+pub mod retry;
 pub mod stream;
+mod tool_loop;
 pub mod types;
 
 use futures::Stream;
@@ -48,71 +52,16 @@ use std::time::Duration;
 
 pub use config::{ApiKey, Config};
 pub use error::{ClientError, ClientResult};
+pub use request::{CreateMessageRequest, Metadata, Speed, ThinkingConfig};
+pub use response::{
+    ApiError, ContentBlockStart, ContentDelta, MessageDelta, MessageResponse, MessageStart,
+    StreamEvent, Usage,
+};
+pub use retry::RetryConfig;
 pub use stream::{EventStream, SseEvent, SseStream};
 pub use types::{
-    ContentBlock, CreateMessageRequest, ExtraToolSchema, Message, MessageResponse, Role, Speed,
-    StreamEvent, ThinkingConfig, ToolChoice, ToolDefinition, Usage,
+    ContentBlock, ExtraToolSchema, Message, MessageContent, Role, ToolChoice, ToolDefinition,
 };
-
-/// Configuration for retry behavior
-#[derive(Debug, Clone)]
-pub struct RetryConfig {
-    /// Maximum number of retries (default: 3)
-    pub max_retries: u32,
-    /// Initial delay before first retry (default: 1s)
-    pub initial_delay: Duration,
-    /// Maximum delay between retries (default: 30s)
-    pub max_delay: Duration,
-    /// Jitter factor for randomizing delays (0.0 to 1.0, default: 0.1)
-    /// A value of 0.1 means delays can vary by up to 10%
-    pub jitter_factor: f64,
-}
-
-impl Default for RetryConfig {
-    fn default() -> Self {
-        Self {
-            max_retries: 3,
-            initial_delay: Duration::from_secs(1),
-            max_delay: Duration::from_secs(30),
-            jitter_factor: 0.1,
-        }
-    }
-}
-
-impl RetryConfig {
-    /// Calculate the delay for a given retry attempt with exponential backoff and jitter.
-    ///
-    /// The delay is calculated as: `initial_delay * 2^attempt * (1.0 - jitter_factor * random)`
-    /// The result is capped at `max_delay`.
-    ///
-    /// # Arguments
-    /// * `attempt` - The retry attempt number (0-indexed)
-    ///
-    /// # Returns
-    /// The calculated delay duration
-    pub fn calculate_delay(&self, attempt: u32) -> Duration {
-        // Calculate base exponential backoff: initial_delay * 2^attempt
-        let base_delay_secs = self.initial_delay.as_secs_f64() * 2_f64.powi(attempt as i32);
-
-        // Cap at max_delay before applying jitter
-        let capped_delay_secs = base_delay_secs.min(self.max_delay.as_secs_f64());
-
-        // Apply jitter: reduce delay by random factor between 0 and jitter_factor
-        // This helps prevent thundering herd when multiple clients retry simultaneously
-        let jitter = self.jitter_factor * random_factor();
-        let jittered_delay_secs = capped_delay_secs * (1.0 - jitter);
-
-        // Ensure we don't go below a minimum delay of 10ms
-        let final_delay_secs = jittered_delay_secs.max(0.01);
-
-        Duration::from_secs_f64(final_delay_secs)
-    }
-}
-
-/// Generate a random factor between 0.0 and 1.0 for jitter calculation.
-fn random_factor() -> f64 {
-    rand::random::<f64>()
-}
 
 /// Anthropic API client
 pub struct Client {
@@ -335,7 +284,7 @@ impl Client {
                 while let Some(result) = stream.next().await {
                     match result {
                         Ok(StreamEvent::ContentBlockDelta {
-                            delta: types::ContentDelta::TextDelta { text },
+                            delta: ContentDelta::TextDelta { text },
                             ..
                         }) => {
                             return Some((Ok(text), stream));
@@ -355,131 +304,6 @@ impl Client {
                 None
             },
         ))
-    }
-
-    /// Execute a message with automatic tool calling loop
-    ///
-    /// This method handles the full tool use protocol:
-    /// 1. Sends initial request with tools
-    /// 2. If Claude returns tool_use blocks, executes them
-    /// 3. Sends tool results back to Claude
-    /// 4. Repeats until Claude returns a text response
-    ///
-    /// # Arguments
-    ///
-    /// * `request` - Initial request with tools configured
-    /// * `tool_executor` - Callback to execute tool calls
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use rustyclawd_core::client::{Client, Config, CreateMessageRequest, Message};
-    ///
-    /// async fn example() -> Result<(), Box<dyn std::error::Error>> {
-    ///     let config = Config::from_default_location().await?;
-    ///     let client = Client::new(config)?;
-    ///
-    ///     let request = CreateMessageRequest::new(
-    ///         "claude-sonnet-4-5-20250929",
-    ///         vec![Message::user("Run ls command")],
-    ///         4096,
-    ///     );
-    ///
-    ///     let response = client.execute_with_tools(request, |tool_name, tool_input| async move {
-    ///         // Execute tool and return result as JSON
-    ///         Ok(serde_json::json!({"output": "file1.txt\nfile2.txt"}))
-    ///     }).await?;
-    ///
-    ///     Ok(())
-    /// }
-    /// ```
-    pub async fn execute_with_tools<F, Fut>(
-        &self,
-        mut request: CreateMessageRequest,
-        tool_executor: F,
-    ) -> ClientResult<MessageResponse>
-    where
-        F: Fn(String, serde_json::Value) -> Fut,
-        Fut: std::future::Future<Output = ClientResult<serde_json::Value>>,
-    {
-        // High limit for complex agentic workflows
-        const MAX_ITERATIONS: usize = 10_000;
-        let mut iteration = 0;
-
-        loop {
-            iteration += 1;
-            if iteration > MAX_ITERATIONS {
-                return Err(ClientError::Unknown(
-                    "Tool execution exceeded maximum iterations".to_string(),
-                ));
-            }
-
-            // Execute the request
-            let response = self.create_message(request.clone()).await?;
-
-            // Check if response contains tool use
-            let mut has_tool_use = false;
-            let mut tool_result_blocks = Vec::new();
-
-            for block in &response.content {
-                if let ContentBlock::ToolUse { id, name, input } = block {
-                    has_tool_use = true;
-
-                    // Log tool invocation details
-                    tracing::debug!(tool = %name, "Invoking tool");
-                    if let Ok(pretty_input) = serde_json::to_string_pretty(input) {
-                        tracing::debug!(tool = %name, input = %pretty_input, "Tool input");
-                    }
-
-                    // Execute the tool
-                    match tool_executor(name.clone(), input.clone()).await {
-                        Ok(result) => {
-                            // Log tool result
-                            if let Ok(pretty_result) = serde_json::to_string_pretty(&result) {
-                                tracing::debug!(tool = %name, result = %pretty_result, "Tool result");
-                            }
-
-                            tool_result_blocks.push(ContentBlock::ToolResult {
-                                tool_use_id: id.clone(),
-                                content: vec![ContentBlock::Text {
-                                    text: result.to_string(),
-                                }],
-                                is_error: None,
-                            });
-                        }
-                        Err(e) => {
-                            // Log tool error
-                            tracing::warn!(tool = %name, error = %e, "Tool execution failed");
-
-                            tool_result_blocks.push(ContentBlock::ToolResult {
-                                tool_use_id: id.clone(),
-                                content: vec![ContentBlock::Text {
-                                    text: format!("Tool execution error: {}", e),
-                                }],
-                                is_error: Some(true),
-                            });
-                        }
-                    }
-                }
-            }
-
-            // If no tool use, we're done
-            if !has_tool_use {
-                return Ok(response);
-            }
-
-            // Build the next request with tool results
-            // First, add the assistant's response with tool_use blocks to conversation
-            request.messages.push(Message::with_blocks(
-                Role::Assistant,
-                response.content.clone(),
-            ));
-
-            // Then add tool results as user message with tool_result blocks
-            request
-                .messages
-                .push(Message::with_blocks(Role::User, tool_result_blocks));
-        }
     }
 
     /// Get the API URL for custom request handling
